RecipeManagementApi/requirements.txt
asgiref==3.11.1
Django==6.0.2
django-filter==25.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
mysqlclient==2.2.8
sqlparse==0.5.5
